pyproject.toml
[tool.poetry]
name = "booklist"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.6"
django-extensions = "^2.2.9"
django-debug-toolbar = "^2.2"
werkzeug = "^1.0.1"
django-environ = "^0.4.5"
djangorestframework = "^3.11.0"
markdown = "^3.2.2"
django-filter = "^2.3.0"
pytest = "^5.4.3"
djoser = "^2.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
